refactor(cookieConsent): extract helpers to src/utils/cookieConsent and update imports
